Reorganize direct orchestration models and create new models
