+++ src/dotnet/Common/Models/Orchestration/Direct/AssistantInputMessage.cs
@@ -1,0 +1,25 @@
+﻿using FoundationaLLM.Common.Constants;
+using System.Text.Json.Serialization;
+
+namespace FoundationaLLM.Common.Models.Orchestration.Direct
+{
+    /// <summary>
+    /// Assistant input message for a direct orchestration request.
+    /// </summary>
+    public class AssistantInputMessage : InputMessage
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AssistantInputMessage"/> class.
+        /// </summary>
+        public AssistantInputMessage() =>
+            Role = InputMessageRoles.Assistant;
+
+        /// <summary>
+        /// An optional name for the participant. Provides the model information
+        /// to differentiate between participants of the same role.
+        /// </summary>
+        [JsonPropertyName("name")]
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+        public string? Name { get; set; }
+    }
+}
