+++ src/dotnet/Common/Models/Orchestration/Direct/AzureAIDirectResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace FoundationaLLM.Common.Models.Orchestration
+namespace FoundationaLLM.Common.Models.Orchestration.Direct
 {
     /// <summary>
     /// The response from the Azure AI orchestration service.
@@ -11,6 +11,6 @@ namespace FoundationaLLM.Common.Models.Orchestration
         /// The completion output from an Azure AI model.
         /// </summary>
         [JsonPropertyName("output")]
-        public string? Output { get; set; } 
+        public string? Output { get; set; }
     }
 }
