--- src/dotnet/Common/Models/Orchestration/DirectCompletionRequest.cs
@@ -1,11 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace FoundationaLLM.Common.Models.Orchestration
-{
-    /// <summary>
-    /// Completion request object that excludes session-based properties.
-    /// </summary>
-    public class DirectCompletionRequest : OrchestrationRequest
-    {
-    }
-}
