+++ src/dotnet/Common/Models/Orchestration/LLMOrchestrationService.cs
@@ -6,15 +6,19 @@
     public enum LLMOrchestrationService
     {
         /// <summary>
-        /// Azure AI orchestration service
+        /// Azure AI orchestration service.
         /// </summary>
         AzureAIDirect,
         /// <summary>
-        /// LangChain orchestration service
+        /// Azure OpenAI orchestration service.
+        /// </summary>
+        AzureOpenAIDirect,
+        /// <summary>
+        /// LangChain orchestration service.
         /// </summary>
         LangChain,
         /// <summary>
-        /// SemanticKernel orchestration service
+        /// SemanticKernel orchestration service.
         /// </summary>
         SemanticKernel
     }
